+++ pkgs/test/lib/src/runner/browser/default_settings.dart
@@ -24,7 +24,11 @@ final defaultSettings = UnmodifiableMapView({
   ),
   Runtime.firefox: ExecutableSettings(
       linuxExecutable: 'firefox',
-      macOSExecutable: '/Applications/Firefox.app/Contents/MacOS/firefox-bin',
+      macOSExecutables: [
+        '/Applications/Firefox.app/Contents/MacOS/firefox-bin',
+        '/Applications/Firefox.app/Contents/MacOS/firefox',
+        'firefox',
+      ],
       windowsExecutable: r'Mozilla Firefox\firefox.exe',
       environmentOverride: 'FIREFOX_EXECUTABLE'),
   Runtime.safari: ExecutableSettings(